+++ nextjs_frontend/src/app/auth/api/authOptions.ts
@@ -1,0 +1,13 @@
+import NextAuth from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+
+export const authOptions = {
+  providers: [
+    GoogleProvider({
+      clientId: process.env.GOOGLE_ID || "",
+      clientSecret: process.env.GOOGLE_SECRET || "",
+    }),
+  ],
+};
+
+export default NextAuth(authOptions);
